fixed movement WASD works (need to fix player too fast when holding down 2 keys at once or at a wall) rendering is good

+++ inc/cub3d_structs.h
@@ -6,7 +6,7 @@
 /*   By: mstefano <mstefano@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/30 20:09:48 by ahusic            #+#    #+#             */
-/*   Updated: 2025/02/01 17:59:33 by mstefano         ###   ########.fr       */
+/*   Updated: 2025/02/01 20:28:47 by mstefano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,15 +20,15 @@
 # define SPEED 5
 # define ROTATION_SPEED 0.045
 # define PLAYER_SPEED 4
-# define NORTH 4.7123889804 // 3 * PI / 2
-# define SOUTH 1.5707963268 // PI / 2
-# define WEST 3.1415926535 // PI
+# define NORTH 4.7123889804
+# define SOUTH 1.5707963268
+# define WEST 3.1415926535
 # define EAST 0
 
 typedef struct s_player
 {
-	int		player_x;
-	int		player_y;
+	double	player_x;
+	double	player_y;
 	double	angle;
 	float	fov;
 	int		rotation;
